Adding a way to globally manage what's open
